Moves AnsiRenderer into a new namespace
