+++ src/markdown-console/Morello.MarkdownConsole/Renderers/AnsiRenderer.Links.cs
@@ -2,7 +2,7 @@ using Markdig.Syntax;
 using Markdig.Syntax.Inlines;
 using Spectre.Console;
 
-namespace Morello.Markdown.Console;
+namespace Morello.Markdown.Console.Renderers;
 
 public partial class AnsiRenderer
 {
