+++ src/markdown-console/Morello.MarkdownConsole/Renderers/AnsiRenderer.Lists.cs
@@ -2,7 +2,7 @@ using Markdig.Extensions.TaskLists;
 using Markdig.Syntax;
 using Spectre.Console;
 
-namespace Morello.Markdown.Console;
+namespace Morello.Markdown.Console.Renderers;
 
 public partial class AnsiRenderer
 {
