+++ src/markdown-console/Morello.MarkdownConsole/Renderers/AnsiRenderer.Tables.cs
@@ -2,7 +2,7 @@ using Markdig.Syntax;
 using Spectre.Console;
 using MarkdownTable = Markdig.Extensions.Tables;
 
-namespace Morello.Markdown.Console;
+namespace Morello.Markdown.Console.Renderers;
 
 public partial class AnsiRenderer
 {
